Handle string quote characters
If the directive is wrapped in string quotes, escape the quote character
on line to transform. Prior this change, the line below would be
incorrectly transformed, using the default pattern:

	Before		After
	strlen("x")	printf("strlen("x")=%d\n", strlen("x"));

Notice the unescaped string quotes in the argument to strlen. After this
change:

	Before		After
	strlen("x")	printf("strlen(\"x\")=%d\n", strlen("x"));

This also works for single quotes (') characters.